Extract test for translating the lexer into its own test file
This is so that you can run a smaller set of test easier when only changing the lexer, or the parser.

This also makes it possible to run each test file in parallel using xargs or parallel when speeding up the test suit.
